controllers: extract limitador controller tests
The idea is to have a test file for each controller.
